+++ screennameviewer/compose/src/main/java/com/donglab/screennameviewer/overlay/renderer/ScreenNameOverlayRenderer.kt
@@ -7,7 +7,7 @@ import android.widget.FrameLayout
 import android.widget.LinearLayout
 import android.widget.TextView
 import com.donglab.screennameviewer.config.ScreenNameOverlayConfig
-import com.donglab.screennameviewer.config.ScreenNameViewerConstants
+import com.donglab.screennameviewer.consts.ScreenNameViewerConstants
 import com.donglab.screennameviewer.overlay.builder.OverlayLayoutBuilder
 import com.donglab.screennameviewer.overlay.builder.StyledTextViewBuilder
 import com.donglab.screennameviewer.util.dp
